+++ Imperium/src/Interface/LayerSelector/LayerSelector.cs
@@ -191,7 +191,8 @@ internal class LayerSelector : BaseUI
 
     private void OnLayerSelect()
     {
-        GameUtils.PlayClip(ImpAssets.GrassClick);
+        if (Imperium.Settings.Preferences.PlaySounds.Value) GameUtils.PlayClip(ImpAssets.GrassClick);
+
         var newMask = ImpUtils.ToggleLayerInMask(layerMaskBinding.Value, selectedLayer);
         layerMaskBinding.Set(newMask);
         layerToggles[selectedLayer].UpdateIsOn(newMask);